FINN-13: Version control the installed XRT version.
Docker installs XRT when building the container. This patch
allows the maintainer to set the Vitis tools version and the
hash of the supported XRT version verified to work with FINN.
This is the way XRT versioning is managed in AWS instances.

FINN-13 was caused by using 2022.1 Vitis and an old XRT version
in usage during 2020.1. Version controlling the sets of tools
will prevent bugs like this from recurring.

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>