[hotfix] mark build dataflow test as xfail
Changes were made to a dataflow step previously that causes the test of the dataflow step to fail.
Further investigation into the options is needed, until then the test is expected to fail.
An option may need to be added to src/finn/qnn-data/build_dataflow/dataflow_build_config.json in order for the test to pass again
but more investigation is needed to determine if that is the best option.
The commit in question is: 59b19dd3699426a549b25f3926716278742ad72b

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>