Add pytest library version and associates plugins for HTML report capturing
Newer version of pytest caused an issue when gathering HTML reports:
ModuleNotFoundError: No module named 'py.xml'; 'py' is not a package

Apparently this is not a pytest bug but due to a related plugin and is caused by depending on the py package but not declaring it as a dependency.
The exact versions of the libraries specified in this commit allow for HTML report gathering.
This was tested in docker and on hardware in 2 virtual environments: virtual_env and conda (zynq and alveo environments respectively)

Signed-off-by: Fionn O'Donohoe <fionno@xilinx.com>